vehicles are creating successfully

+++ src/config/db.ts
@@ -7,20 +7,20 @@ export const pool = new Pool({
 
 const initDB = async () => {
     await pool.query(`
-    CREATE TABLE IF NOT EXISTS vehicles (
+CREATE TABLE IF NOT EXISTS vehicles (
     id SERIAL PRIMARY KEY,
     vehicle_name VARCHAR(100) NOT NULL,
     type VARCHAR(50) NOT NULL
         CHECK (type IN ('car', 'bike', 'van', 'suv'))
         CHECK (type = LOWER(type)),
     registration_number VARCHAR(150) UNIQUE NOT NULL,
-    daily_rent_price NUMERIC(10,2) NOT NULL
-        CHECK (daily_rent_price > 0),
+    daily_rent_price INTEGER NOT NULL CHECK (daily_rent_price > 0),
     availability_status VARCHAR(50) NOT NULL
         CHECK (availability_status IN ('available', 'booked'))
         CHECK (availability_status = LOWER(availability_status))
 );
 `);
+
 };
